+++ dataset_update.go
@@ -3,7 +3,6 @@ package goqu
 import (
 	"reflect"
 	"sort"
-	"time"
 )
 
 func (me *Dataset) canUpdateField(field reflect.StructField) bool {
@@ -80,9 +79,8 @@ func (me *Dataset) ToUpdateSql(update interface{}) (string, []interface{}, error
 func (me *Dataset) getUpdateExpressions(value reflect.Value) (updates []UpdateExpression) {
 	for i := 0; i < value.NumField(); i++ {
 		v := value.Field(i)
-		kind := v.Kind()
-		if reflect.TypeOf(v.Interface()).Name() == reflect.TypeOf((*time.Time)(nil)).Elem().Name() || (kind != reflect.Struct && kind != reflect.Ptr) {
-			t := value.Type().Field(i)
+		t := value.Type().Field(i)
+		if !t.Anonymous {
 			if me.canUpdateField(t) {
 				updates = append(updates, I(t.Tag.Get("db")).Set(v.Interface()))
 			}